pyproject.toml
[tool.poetry]
name = "polyglottal_backend"
version = "0.1.0"
description = ""
authors = ["hono1030 <honoka.n1030@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
start = "polyglottal_backend.main:start"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
pydantic = "^2.9.2"
python-dotenv = "^1.0.1"
pymongo = {extras = ["srv"], version = "^4.10.1"}
dnspython = "^2.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
